update and document history apis
